A refactoring related change in the BPMN library caused the imports to fail.
