+++ eclipse/org.eclipse.bpmn2.modeler.runtime.marklogic/src/org/eclipse/bpmn2/modeler/runtime/marklogic/workflow/MarkLogicWorkflowRuntimeExtension.java
@@ -48,7 +48,7 @@ import org.eclipse.bpmn2.modeler.core.LifecycleEvent.EventType;
 //import org.eclipse.bpmn2.modeler.core.runtime.TargetRuntime;
 import org.eclipse.bpmn2.modeler.core.utils.ModelUtil.Bpmn2DiagramType;
 
-import org.eclipse.bpmn2.modeler.ui.DefaultBpmn2RuntimeExtension.RootElementParser;
+import org.eclipse.bpmn2.modeler.ui.AbstractBpmn2RuntimeExtension.RootElementParser;
 import org.eclipse.bpmn2.modeler.ui.editor.BPMN2Editor;
 import org.eclipse.bpmn2.modeler.ui.wizards.FileService;
 import org.eclipse.core.resources.IContainer;
